quicknotes_api/pyproject.toml
[project]
name = "quicknotes-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "faker>=38.2.0",
    "python-dotenv>=1.2.1",
]
